# docs/swagger.yaml
basePath: /
definitions:
...
  domain.CreateWeaponRequest:
    properties:
      content:
        type: string
      level:
        type: integer
      name:
        type: string
      story:
        type: string
      type:
        type: integer
    type: object
...
  domain.UpdateWeaponRequest:
    properties:
      content:
        type: string
      level:
        type: integer
      name:
        type: string
      story:
        type: string
      type:
        type: integer
    type: object
...
  model.Weapon:
    description: 武器信息
    properties:
      content:
        example: AK47 is a popular weapon in the world
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      level:
        example: 1
        type: integer
      name:
        example: AK47
        type: string
      story:
        example: AK47 is a popular weapon in the world
        type: string
      type:
        example: 1
        type: integer
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  pagination.PageResponse:
    properties:
      data:
        description: 数据列表
      has_next:
        description: 是否有下一页
        type: boolean
      has_prev:
        description: 是否有上一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页大小
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
        type: string
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a dove server API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: dove API
  version: "1.0"
paths:
...
  /weapons/:
    get:
      consumes:
      - application/json
      description: 获取所有Weapon列表（支持分页、排序和搜索）
      parameters:
      - description: 页码，默认为1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页大小，默认为10，最大100
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: '排序字段：ID, Name, '
        in: query
        name: sort_by
        type: string
      - description: 排序方向：asc, desc，默认为desc
        in: query
        name: sort_order
        type: string
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 搜索字段：Name, ，不指定则在所有字段中搜索
        in: query
        name: search_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/pagination.PageResponse'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/model.Weapon'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取所有Weapon
      tags:
      - 武器
    post:
      consumes:
      - application/json
      description: 创建新的Weapon
      parameters:
      - description: Weapon信息
        in: body
        name: weapon
        required: true
        schema:
          $ref: '#/definitions/domain.CreateWeaponRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Weapon'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建Weapon
      tags:
      - 武器
  /weapons/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除Weapon
      parameters:
      - description: WeaponID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除Weapon
      tags:
      - 武器
    get:
      consumes:
      - application/json
      description: 根据ID获取Weapon详细信息
      parameters:
      - description: WeaponID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Weapon'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取Weapon详情
      tags:
      - 武器
    put:
      consumes:
      - application/json
      description: 根据ID更新Weapon信息
      parameters:
      - description: WeaponID
        in: path
        name: id
        required: true
        type: integer
      - description: Weapon更新信息
        in: body
        name: weapon
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateWeaponRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Weapon'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新Weapon
      tags:
      - 武器
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"